Refactor of tests samples, call for test using a file instead a directory
