+++ src/test/java/com/gdssecurity/pmd/CWE113HttpResponseSplittingTest.java
@@ -11,13 +11,13 @@ public class CWE113HttpResponseSplittingTest {
 
 	@Test
 	public void testUnvalidatedRedirects() throws Exception {
-		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe113responsesplitting/redirect",
+		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe113responsesplitting/TestRedirectServlet.java",
 				PMDRunner.RULESET_UNVALIDATED_REDIRECTS));
 	}
 	
 	@Test
 	public void testHTTPResponseSplitting() throws Exception {
-		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe113responsesplitting/redirect",
+		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe113responsesplitting/TestRedirectServlet.java",
 				PMDRunner.RULESET_HTTP_RESPONSE_SPLITTING));
 	}
 }
