+++ src/test/java/com/gdssecurity/pmd/CWE328ReversibleHashTest.java
@@ -12,6 +12,6 @@ public class CWE328ReversibleHashTest {
 	@Test
 	public void test1() throws Exception {
 		Assert.assertEquals(1,
-				PMDRunner.run("src/test/java/resources/cwe328reversiblehash", PMDRunner.RULESET_VULNERABLE));
+				PMDRunner.run("src/test/java/resources/cwe328reversiblehash/CWE328_Reversible_One_Way_Hash__basic_01.java", PMDRunner.RULESET_VULNERABLE));
 	}
 }
