+++ src/test/java/com/gdssecurity/pmd/CWE931XssTest.java
@@ -13,12 +13,13 @@ public class CWE931XssTest {
 
 	@Test
 	public void test1() throws Exception {
-		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe931xss/xss01", PMDRunner.RULESET_XSS));
+		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe931xss/TestXSSServlet.java", PMDRunner.RULESET_XSS));
 	}
 
 	@Test
 	public void test2() throws Exception {
-		Assert.assertEquals(2, PMDRunner.run("src/test/java/resources/cwe931xss/xss02jsp", PMDRunner.RULESET_XSS));
+		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe931xss/XSS1_jsp.java", PMDRunner.RULESET_XSS));
+		Assert.assertEquals(1, PMDRunner.run("src/test/java/resources/cwe931xss/XSS2_jsp.java", PMDRunner.RULESET_XSS));
 	}
 
 }
