+++ src/test/java/resources/cwe931xss/XSS2_jsp.java
@@ -1,4 +1,4 @@
-package resources.cwe931xss.xss02jsp;
+package resources.cwe931xss;
 
 import java.util.List;
